requirements.txt
certifi==2019.3.9
chardet==3.0.4
da-vinci==0.2.2
Django==2.2.4
django-crispy-forms==1.7.2
django-thumbnails==0.2.1
gunicorn==19.9.0
idna==2.8
Pillow==6.1.0
pytz==2019.1
requests==2.22.0
shortuuid==0.5.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
